refactor: separate GraduationProgressContent and fix build issues
- Extract GraduationProgressContent component to features/academic
- Separate GraduationNavigationHeader with proper Suspense handling
- Remove cache: 'no-store' to prevent static generation issues
- Update useGraduationNavigation to work with useSuspenseQuery
- Add force-dynamic export to graduation progress layout
