+++ src/app/(main)/graduation-progress/layout.tsx
@@ -1,31 +1,7 @@
-'use client';
-
-import { TopNavigation } from '@/components/ui/TopNavigation';
-import { useGraduationNavigation } from '@/features/academic/hooks/useGraduationNavigation';
+import GraduationNavigationHeader from '@/features/academic/components/progress/GraduationNavigationHeader';
 import styles from './layout.module.scss';
 
-function GraduationNavigationHeader() {
-  const { navigationTitle, handleBack, isLoading } = useGraduationNavigation();
-
-  if (isLoading) {
-    return (
-      <TopNavigation.Preset 
-        title="졸업요건" 
-        type="back" 
-        onNavigationClick={handleBack} 
-      />
-    );
-  }
-
-  return (
-    <TopNavigation.Preset 
-      title={navigationTitle} 
-      type="back" 
-      onNavigationClick={handleBack} 
-    />
-  );
-}
-
+export const dynamic = 'force-dynamic';
 export default function GraduationProgressLayout({ children }: { children: React.ReactNode }) {
   return (
     <div className={styles.container}>
